+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to
 
 ### Changed
 
+- Normalize the organization code field
 - Activate pagination by default on all endpoints (20 items per page)
 - Use user fullname instead of username in order confirmation email
 - Rename certificate field into certificate_definition for the ProductSerializer